+++ sdk/appconfiguration/azure-appconfiguration/CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Bugs Fixed
 
 ### Other Changes
-- Bumped minimum dependency on `azure-core` to `>=1.28.0`.
+- Bumped minimum dependency on `azure-core` to `1.28.0`.
 - Updated the default `api_version` to "2023-10-01".
 
 ## 1.5.0b3 (2023-10-10)